Revert "Parity refactor (#68)"
This reverts commit 3f82524736ea5c508b777024cd4c0522774d6fe0.
