+++ HISTORY.md
@@ -1,16 +1,4 @@
-2.11.0 / 2017-11-07
-=================
-
-* Pulled out classic, universal, and enhanced-ecommerce support into their own files and required them into index.js
-* Adds test improvments.
-* Fixes bug in initialize function.
-* Removed some repetitive code by refactoring out into helper functions.
-* Added some improved commenting
-* Made our mapping logic a little more clear for anyone reading the code
-* Removed the need to reference our methods to methods with enhanced or classic appended to them.
-* Got rid of the behavior where we were passing undefined values as params to ga methods for easier testing.
-
-2.9.5 / 2017-10-26
+2.9.4 / 2017-10-26
 =================
 
 * Release to support reversion of failed deploy.
