--- lib/classic.js
@@ -1,164 +1,0 @@
-'use strict';
-
-/**
- * Module dependencies.
- */
-
-var Track = require('segmentio-facade').Track;
-var each = require('component-each');
-var is = require('is');
-var push = require('global-queue')('_gaq');
-var useHttps = require('use-https');
-var formatValue = require('./utils').formatValue;
-var path = require('./utils').path;
-
-
-/**
- * Page (classic).
- *
- * https://developers.google.com/analytics/devguides/collection/gajs/methods/gaJSApiBasicConfiguration
- *
- * @param {Page} page
- */
-
-exports.page = function(page) {
-  var category = page.category();
-  var props = page.properties();
-  var name = page.fullName();
-  var track;
-
-  push('_trackPageview', path(props, this.options));
-
-  // categorized pages
-  if (category && this.options.trackCategorizedPages) {
-    track = page.track(category);
-    this.track(track, { nonInteraction: 1 });
-  }
-
-  // named pages
-  if (name && this.options.trackNamedPages) {
-    track = page.track(name);
-    this.track(track, { nonInteraction: 1 });
-  }
-};
-
-/**
- * Loaded? (classic)
- *
- * @return {Boolean}
- */
-
-exports.loaded = function() {
-  return !!(window._gaq && window._gaq.push !== Array.prototype.push);
-};
-
-/**
- * Initialize (classic).
- *
- * https://developers.google.com/analytics/devguides/collection/gajs/methods/gaJSApiBasicConfiguration
- */
-
-exports.initialize = function() {
-  var opts = this.options;
-  var anonymize = opts.anonymizeIp;
-  var domain = opts.domain;
-  var enhanced = opts.enhancedLinkAttribution;
-  var ignore = opts.ignoredReferrers;
-  var sample = opts.siteSpeedSampleRate;
-
-  window._gaq = window._gaq || [];
-  push('_setAccount', opts.trackingId);
-  push('_setAllowLinker', true);
-
-  if (anonymize) push('_gat._anonymizeIp');
-  if (domain) push('_setDomainName', domain);
-  if (sample) push('_setSiteSpeedSampleRate', sample);
-
-  if (enhanced) {
-    var protocol = document.location.protocol === 'https:' ? 'https:' : 'http:';
-    var pluginUrl = protocol + '//www.google-analytics.com/plugins/ga/inpage_linkid.js';
-    push('_require', 'inpage_linkid', pluginUrl);
-  }
-
-  if (ignore) {
-    if (!is.array(ignore)) ignore = [ignore];
-    each(ignore, function(domain) {
-      push('_addIgnoredRef', domain);
-    });
-  }
-
-  if (this.options.doubleClick) {
-    this.load('double click', this.ready);
-  } else {
-    var name = useHttps() ? 'https' : 'http';
-    this.load(name, this.ready);
-  }
-};
-
-/**
- * Track (classic).
- *
- * https://developers.google.com/analytics/devguides/collection/gajs/methods/gaJSApiEventTracking
- *
- * @param {Track} track
- */
-
-exports.track = function(track, options) {
-  var opts = options || track.options(this.name);
-  var props = track.properties();
-  var revenue = track.revenue();
-  var event = track.event();
-  var category = this._category || track.category() || 'All';
-  var label = props.label;
-  var value = formatValue(revenue || props.value);
-  var nonInteraction = !!(props.nonInteraction || opts.nonInteraction);
-  push('_trackEvent', category, event, label, value, nonInteraction);
-};
-
-/**
- * Completed order.
- *
- * https://developers.google.com/analytics/devguides/collection/gajs/gaTrackingEcommerce
- * https://developers.google.com/analytics/devguides/collection/gajs/gaTrackingEcommerce#localcurrencies
- *
- * @param {Track} track
- * @api private
- */
-
-exports.orderCompleted = function(track) {
-  var total = track.total() || track.revenue() || 0;
-  var orderId = track.orderId();
-  var products = track.products() || [];
-  var props = track.properties();
-  var currency = track.currency();
-
-  // required
-  if (!orderId) return;
-
-  // add transaction
-  push('_addTrans',
-    orderId,
-    props.affiliation,
-    total,
-    track.tax(),
-    track.shipping(),
-    track.city(),
-    track.state(),
-    track.country());
-
-  // add items
-  each(products, function(product) {
-    var track = new Track({ properties: product });
-    push('_addItem',
-      orderId,
-      track.sku(),
-      track.name(),
-      track.category(),
-      track.price(),
-      track.quantity());
-  });
-
-  // send
-  push('_set', 'currencyCode', currency);
-  push('_trackTrans');
-};
