+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@segment/analytics.js-integration-google-analytics",
   "description": "The Google Analytics analytics.js integration.",
-  "version": "2.11.0",
+  "version": "2.9.5",
   "keywords": [
     "analytics.js",
     "analytics.js-integration",
@@ -24,7 +24,6 @@
   "homepage": "https://github.com/segment-integrations/analytics.js-integration-google-analytics#readme",
   "dependencies": {
     "@ndhoule/defaults": "^2.0.1",
-    "@ndhoule/extend": "^2.0.0",
     "@segment/analytics.js-integration": "^3.1.0",
     "component-each": "^0.2.6",
     "global-queue": "^1.0.1",
@@ -58,7 +57,6 @@
     "mocha": "^2.2.5",
     "npm-check": "^5.2.1",
     "phantomjs-prebuilt": "^2.1.7",
-    "sinon": "^4.0.1",
     "to-array": "^0.1.4",
     "watchify": "^3.7.0"
   }
